Moved RequestMapping Shortcut annotations article code from spring-mvc-java to spring-mvc-basics
